requirements.txt
easyocr==1.7.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
pdf2image==1.16.3
scikit-image==0.21.0
imutils==0.5.4
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyYAML

# Core Dependencies
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80
pdf2image==1.17.0

# OCR Engine
paddleocr==2.7.3
paddlepaddle==2.6.1

# YAML Configuration
PyYAML==6.0.1

# Utilities
numpy==1.26.4


